refactor: update eventType filename
