+++ .github/whitelist-capital-letters.txt
@@ -45,7 +45,6 @@
 ./libs/common/src/enums/notificationType.ts
 ./libs/common/src/enums/keySuffixOptions.ts
 ./libs/common/src/enums/productType.ts
-./libs/common/src/enums/eventType.ts
 ./libs/common/src/enums/hashPurpose.ts
 ./libs/common/src/enums/uriMatchType.ts
 ./libs/common/src/enums/secureNoteType.ts
