+++ libs/common/src/abstractions/event/event-collection.service.ts
@@ -1,4 +1,4 @@
-import { EventType } from "../../enums/eventType";
+import { EventType } from "../../enums/event-type";
 
 export abstract class EventCollectionService {
   collect: (
