+++ libs/common/src/models/view/event.view.ts
@@ -1,5 +1,5 @@
 import { EventSystemUser } from "../../enums/event-system-user";
-import { EventType } from "../../enums/eventType";
+import { EventType } from "../../enums/event-type";
 
 export class EventView {
   message: string;
